pyproject.toml
[tool.poetry]
name = "brother_label_printer_control"
version = "0.8.0"
description = "A library for creating and printing labels for Brother P-Touch devices"
authors = [
    "Adrian Tschira <packages@notafile.com>",
    "Benedikt Neuffer <benedikt.neuffer@kit.edu>",
    "Ben Burwood <ben.burwood@outlook.com>"
]
license = "Apache License, Version 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^9.5.0"
pyusb = "^1.2.1"
packbits = "^0.6"
pyserial = {version = "^3.5", optional = true}
qrcode = {version = "^7.4.2", optional = true, extras = ["pil"]}
setuptools = "^68.0.0"
pysnmp = {version = "^4.4.12", optional = true}
#pin pyasn because pysnmp does not work with newer versions
pyasn1 = {version = "^0.4.8", optional = true}

[tool.poetry.extras]
bluetooth = ["pyserial"]
qrcode = ["qrcode"]
snmp = ["pysnmp", "pyasn1"]

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
types-setuptools = "^68.1.0.0"
types-pillow = "^10.0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
